Fix delay function for lower delays
The debounce feature expects the button to be held for a minimum amount of time
to register a button press. This caused the delay function to always return 0
(no button presses) for very short delays. It will now return 1 for short
delays iff the button was pressed/held during that time.