Automation: Fix for Switch v11 firmware
The latest Switch firmware adds a “Switch Online” button on the main
menu, on the right side of the “News” button. This changes the way the
controllers or console settings can be accessed. To maintain
compatibility with previous software versions, the automation first
sends the cursor to the rightmost position, then backs a couple steps
to position it on the correct button.