Sword/Shield: Handle multiple users at restart
If there are multiple users on the console, when starting a game, a
user selection screen will pop up. To handle this, the automation
nows mashes A when restarting the game, in order to validate the
selection screen (or do nothing if there is no selection screen).

Note that this may not work correctly if the user selected by default
on this screen is not the one that originally started the game.